Added basic rating functionality, created extension method for retrieving user id, and converted helper function to be generic

+++ src/DrinksWeMake.Api/Features/Ratings/DeleteRating.cs
@@ -1,13 +1,17 @@
+using System.Security.Claims;
 using DrinksWeMake.Api.Data;
+using DrinksWeMake.Api.Extensions;
 using Microsoft.EntityFrameworkCore;
 
 namespace DrinksWeMake.Api.Features.Ratings;
 
 public static class DeleteRating
 {
-   private static async Task<IResult> Handle(AppDbContext dbContext, int id, CancellationToken cancellationToken)
+   private static async Task<IResult> Handle(ClaimsPrincipal user, AppDbContext dbContext, int id, CancellationToken cancellationToken)
    {
-      var numDeletedRatings = await dbContext.Ratings.Where(r => r.Id == id).ExecuteDeleteAsync(cancellationToken);
+      var userId = user.GetUserId();
+      
+      var numDeletedRatings = await dbContext.Ratings.Where(r => r.Id == id && r.UserId == userId).ExecuteDeleteAsync(cancellationToken);
 
       return numDeletedRatings == 0 ? Results.NotFound() : Results.NoContent();
    }
